Collapsed ViT components into one file, simplified patch embedding layer
