+++ keras_cv/layers/__init__.py
@@ -74,6 +74,5 @@ from keras_cv.layers.regularization.squeeze_excite import SqueezeAndExcite2D
 from keras_cv.layers.regularization.stochastic_depth import StochasticDepth
 from keras_cv.layers.spatial_pyramid import SpatialPyramidPooling
 from keras_cv.layers.transformer_encoder import TransformerEncoder
-from keras_cv.layers.vit_class_tokenizing import ClassTokenizing
-from keras_cv.layers.vit_patch_embedding import PatchEmbedding
-from keras_cv.layers.vit_patching import Patching
+from keras_cv.layers.vit_layers import PatchEmbedding
+from keras_cv.layers.vit_layers import Patching
